fix: narrow conversion warning in fuzzyScore on 64-bit Linux
Cast qsizetype arithmetic result to int to silence
-Werror=conversion with GCC 13 on Linux (qsizetype is long long).

+++ src/ui/src/commandpalette.cpp
@@ -332,7 +332,7 @@ int CommandPalette::fuzzyScore( const QString& pattern, const QString& text )
     }
 
     // Penalty for remaining unmatched text length
-    score += ( textLen - patternLen ) / 3;
+    score += static_cast<int>( ( textLen - patternLen ) / 3 );
 
     return score;
 }
